+++ tests/unittests/command_parse/test_geo.py
@@ -21,31 +21,3 @@ def test_geoadd(judge_command):
             "member": '"Catania"',
         },
     )
-
-
-def test_georadiusbymember(judge_command):
-    judge_command(
-        "GEORADIUSBYMEMBER Sicily Agrigento 100 km",
-        {
-            "command": "GEORADIUSBYMEMBER",
-            "key": "Sicily",
-            "member": "Agrigento",
-            "float": "100",
-            "distunit": "km",
-        },
-    )
-
-
-def test_georadius(judge_command):
-    judge_command(
-        "GEORADIUS Sicily 15 37 200 km WITHDIST WITHCOORD ",
-        {
-            "command": "GEORADIUS",
-            "key": "Sicily",
-            "longitude": "15",
-            "latitude": "37",
-            "float": "200",
-            "distunit": "km",
-            "geochoice": "WITHCOORD",
-        },
-    )