+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Feature: support new command: `FAILOVER`.
 - Feature: `FLUSHDB` and `FLUSHALL` supports `SYNC` option.
 - Feature: `GEOADD` supports `CH XX NX` options.
+- Deprecate: `GEORADIUS` is deprecated, no auto-complete for this command anymore.
+- Deprecate: `GEORADIUSBYMEMBER` is deprecated, no auto-complete for this command anymore.
 
 ## 1.11.1